build: SVGR 도입 및 PNG→SVG 마이그레이션; refactor(auth): id→email
