+++ vite.config.js
@@ -1,7 +1,12 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import svgr from "vite-plugin-svgr";
 
-// https://vite.dev/config/
 export default defineConfig({
-  plugins: [react()],
-})
+  plugins: [react(), svgr({
+    svgrOptions: {
+      svgo: true,
+      svgoConfig: { plugins: [{ name: "prefixIds" }] },
+    },
+  })],
+});
